+++ pandas/core/sparse/api.py
@@ -1,6 +1,5 @@
-from pandas.core.arrays.sparse import (
-    SparseArray,
-    SparseDtype,
-)
+from pandas.core.dtypes.dtypes import SparseDtype
+
+from pandas.core.arrays.sparse import SparseArray
 
 __all__ = ["SparseArray", "SparseDtype"]
