+++ pandas/tests/arrays/sparse/test_arithmetics.py
@@ -4,11 +4,9 @@ import numpy as np
 import pytest
 
 import pandas as pd
+from pandas import SparseDtype
 import pandas._testing as tm
-from pandas.core.arrays.sparse import (
-    SparseArray,
-    SparseDtype,
-)
+from pandas.core.arrays.sparse import SparseArray
 
 
 @pytest.fixture(params=["integer", "block"])
