+++ pandas/tests/arrays/sparse/test_array.py
@@ -6,12 +6,12 @@ import pytest
 from pandas._libs.sparse import IntIndex
 
 import pandas as pd
-from pandas import isna
-import pandas._testing as tm
-from pandas.core.arrays.sparse import (
-    SparseArray,
+from pandas import (
     SparseDtype,
+    isna,
 )
+import pandas._testing as tm
+from pandas.core.arrays.sparse import SparseArray
 
 
 @pytest.fixture
