+++ pandas/tests/arrays/sparse/test_astype.py
@@ -3,12 +3,12 @@ import pytest
 
 from pandas._libs.sparse import IntIndex
 
-from pandas import Timestamp
-import pandas._testing as tm
-from pandas.core.arrays.sparse import (
-    SparseArray,
+from pandas import (
     SparseDtype,
+    Timestamp,
 )
+import pandas._testing as tm
+from pandas.core.arrays.sparse import SparseArray
 
 
 class TestAstype:
