+++ pandas/tests/reshape/test_get_dummies.py
@@ -13,13 +13,11 @@ from pandas import (
     DataFrame,
     RangeIndex,
     Series,
+    SparseDtype,
     get_dummies,
 )
 import pandas._testing as tm
-from pandas.core.arrays.sparse import (
-    SparseArray,
-    SparseDtype,
-)
+from pandas.core.arrays.sparse import SparseArray
 
 
 class TestGetDummies:
